Add unique function
My unique function works like this: Given an expression like this (A A B C) or (A D D C) with redundant atoms, it reduces it into (A B C) or (A D C), respectively. If the expression contains terms with more redundancy, as in (A A A B), we can get rid of all redundant terms by applying (collapse output) to the function.